# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Dokumentasi API untuk website DP3AP2KB",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/settings": {
            "get": {
                "tags": [
                    "Settings"
                ],
                "summary": "Get site settings",
                "operationId": "9e05b9249b6dde9fbaee4288a0016fb0",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Settings"
                ],
                "summary": "Update site settings (Use _method=PUT)",
                "operationId": "b48b60240382bbe81e894669a6f951b4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "_method"
                                ],
                                "properties": {
                                    "_method": {
                                        "type": "string",
                                        "example": "PUT"
                                    },
                                    "site_name": {
                                        "type": "string"
                                    },
                                    "copyright_text": {
                                        "type": "string"
                                    },
                                    "site_logo": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "footer_about": {
                                        "type": "string"
                                    },
                                    "alamat": {
                                        "type": "string"
                                    },
                                    "jam_kerja[]": {
                                        "description": "Array of strings",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "telepon[]": {
                                        "description": "Array of strings",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "facebook_url": {
                                        "type": "string",
                                        "format": "url"
                                    },
                                    "instagram_url": {
                                        "type": "string",
                                        "format": "url"
                                    },
                                    "twitter_url": {
                                        "type": "string",
                                        "format": "url"
                                    },
                                    "youtube_url": {
                                        "type": "string",
                                        "format": "url"
                                    },
                                    "whatsapp_url": {
                                        "type": "string",
                                        "format": "url"
                                    },
                                    "google_map_url": {
                                        "type": "string"
                                    },
                                    "tiktok_url": {
                                        "type": "string",
                                        "format": "url"
                                    },
                                    "notification_email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Settings updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}